requirements.txt
Django>=5.0
djangorestframework~=3.16.1
psycopg2-binary
langgraph
openai~=1.107.3
redis
django-environ
gunicorn

requests~=2.32.5
bs4~=0.0.2
beautifulsoup4~=4.13.5